Trajectory plotting function and tracker method A new function in plot.py "plot_trajectory" allows the user to display a a trajectory in a single image, over a given background (could be a coastline).

+++ trajectory.py
@@ -39,7 +39,7 @@ class CycloneObject:
         # Erases all pixels that are not part of the object
         mask[np.logical_not(self.image)] = 0
 
-        self.mask = mask
+        self.mask = mask.astype(int)
 
 
 class Trajectory:
